抢单模块除地图功能之外的功能全部完成，包括许多细节

+++ HousekeepingServer/Classes/Main/View/HSTabBar.m
@@ -50,8 +50,8 @@
   if ([self.delegate
           respondsToSelector:@selector(tabBar:didSelectedFrom:to:)]) {
     [self.delegate tabBar:self
-          didSelectedFrom:self.selectedBtn.tag
-                       to:button.tag];
+          didSelectedFrom:(int)self.selectedBtn.tag
+                       to:(int)button.tag];
   }
   self.selectedBtn.selected = NO;
   button.selected = YES;
